Added testing function for various types of honeycombs.
Added wrong test results to honeycomb-sphere.dat to check that the section is deleted.

Added the attribute is_middle_node to Node, so that no point BCs are applied to the middle node of a beam. The effects of this should be checked when a lineload is added to a beam.
Also added it to the Beam elements creation list.

Made the elements per beam in the create_honeycomb functions a keyword argument.

Added read_dat to constructor of InputFile.